+++ src/taphome_sdk/__init__.py
@@ -39,7 +39,13 @@ from .helpers import (
 )
 from .observable import Event, ObservableValue
 from .operation_mode import OperationMode
-from .taphome_api import ApiConnectionType, Location, TapHomeApi
+from .taphome_api import (
+    ApiConnectionType,
+    Location,
+    TapHomeApi,
+    ValueChangeFailedException,
+    ValueChangeResult,
+)
 from .taphome_hub import (
     DeviceNotExposedError,
     DeviceTypeError,