+++ src/taphome_sdk/helpers.py
@@ -4,7 +4,7 @@ import asyncio
 from collections.abc import Awaitable, Callable
 from datetime import timedelta
 from enum import Enum
-from typing import Any, Protocol, TypeVar
+from typing import Any, Protocol, Self, TypeVar
 
 EnumT = TypeVar("EnumT", bound=Enum)
 
@@ -31,7 +31,7 @@ def enum_from_string_required(enum_type: type[EnumT], value: str) -> EnumT:
     return enum
 
 
-def get_required(config: dict[str, Any], key: str):
+def get_required(config: dict[str, Any], key: str) -> Any:
     """Return value for ``key`` or raise if missing."""
     value = get_optional(config, key, None)
     if value is None:
@@ -39,25 +39,25 @@ def get_required(config: dict[str, Any], key: str):
     return value
 
 
-def get_optional(config: dict[str, Any], key: str, default):
+def get_optional(config: dict[str, Any], key: str, default: Any) -> Any:
     """Return value for ``key`` or ``default`` if not present."""
     if isinstance(config, dict) and key in config:
         return config[key]
     return default
 
 
-class FromDictProtocol[ResponseT](Protocol):
+class FromDictProtocol(Protocol):
     """Protocol for classes that can be created from dict."""
 
     @classmethod
-    def from_dict(cls, data: dict) -> ResponseT:
+    def from_dict(cls, data: dict[str, Any]) -> Self:
         """Create an instance of the class from a dictionary."""
         ...
 
 
 def set_interval_async(
     func: Callable[..., Awaitable[Any]], interval: timedelta, *args: Any, **kwargs: Any
-) -> asyncio.Task:
+) -> asyncio.Task[None]:
     """Schedule `func(*args, **kwargs)` every `interval` seconds in asyncio."""
 
     async def periodic_execute() -> None: